build 20220214
There is almost no change in the core.
This version introduces a GRAPHICS.f library that can be loaded via NEEDS GRAPHICS
and makes available PLOT, DRAW-LINE and CIRCLE for *every* Graphic-Mode available
using LAYERx words.
This GRAPHICS libary is a valuable piece of code that is a good example of
vector-call via  DEFER and IS along with the use of DOES> to create very compact
and fast code.